OBMFULL-5394 Implement ehcache statistics API for the crash commands
Conflicts:
	java/sync/opush/push-dao/src/test/java/org/obm/push/store/ehcache/TestingEhCacheConfiguration.java
	java/sync/opush/push-dao-ehcache/src/main/java/org/obm/push/store/ehcache/EhCacheDaoModule.java
	java/sync/opush/push-dao-ehcache/src/main/java/org/obm/push/store/ehcache/FakeEhCacheStatistics.java
	java/sync/opush/push-dao-testsuite/src/main/java/org/obm/push/dao/testsuite/SyncedCollectionDaoTest.java

+++ java/sync/opush/push-dao-ehcache/src/main/java/org/obm/push/store/ehcache/EhCacheStatistics.java
@@ -36,20 +36,20 @@ public interface EhCacheStatistics {
 	/**
 	 * @return disk gets count for a short time before this call
 	 */
-	int shortTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
+	long shortTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
 
 	/**
 	 * @return disk gets count for a medium time before this call
 	 */
-	int mediumTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
+	long mediumTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
 
 	/**
 	 * @return disk gets count for a long time before this call
 	 */
-	int longTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
+	long longTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
 
 	/**
 	 * @return current heap memory took by the given store, in bytes
 	 */
-	int memorySizeInBytes(String storeName);
+	long memorySizeInBytes(String storeName);
 }
